breaks out code into multiple js modules
